Audio
Added AudioStreamPlayer2D for music, make a scene for it and added the scene in Autoload in Project Settings, to keep the music playing, even the scene reload. Also, was added AudioStreamPlayer2D for coins SFX. For play audio when coins are collected, was added a new animation that plays every time that a coin is collected. This animation handle with hide sprite, play the SFX and calls queue_free function to delete the object